+++ README.md
@@ -2,7 +2,7 @@
 
 <a href="https://arxiv.org/abs/2209.15001"><img src="https://img.shields.io/badge/arXiv-Dilated%20Neighborhood%20Attention%20Trasnformer-%23C209C1" /></a>
 <a href="https://arxiv.org/abs/2204.07143"><img src="https://img.shields.io/badge/arXiv-Neighborhood%20Attention%20Trasnformer-%2300B0F0" /></a>
-[<img src="https://img.shields.io/badge/Extension-Neighborhood%20Attention%20CUDA%20Extension%20for%20PyTorch-%23fc6562" />](NATTEN.md)
+[<img src="https://img.shields.io/badge/CUDA%20Extension-NATTEN-%23fc6562" />](https://github.com/SHI-Labs/NATTEN)
 
 ![NAT-Intro](assets/dinat/intro_dark.png#gh-dark-mode-only)
 ![NAT-Intro](assets/dinat/intro_light.png#gh-light-mode-only)
@@ -10,9 +10,10 @@
 **Powerful hierarchical vision transformers based on sliding window attention.**
 
 Neighborhood Attention (NA, local attention) was introduced in our original paper, 
-[NAT](NAT.md), and runs efficiently with our CUDA extension to PyTorch, [NATTEN](NATTEN.md).
+[NAT](NAT.md), and runs efficiently with our CUDA extension to PyTorch, [NATTEN](https://github.com/SHI-Labs/NATTEN).
 
-We recently introduced a new model, [DiNAT](DiNAT.md), which extends NA by dilating neighborhoods (DiNA, sparse global attention).
+We recently introduced a new model, [DiNAT](DiNAT.md), 
+which extends NA by dilating neighborhoods (DiNA, sparse global attention, a.k.a. dilated local attention).
 
 Combinations of NA/DiNA are capable of preserving locality, 
 expanding the receptive field exponentially, 
@@ -46,41 +47,30 @@ Note that the edges are special (edge) cases.
 
 # News
 
+### October 8, 2022
+* [NATTEN](https://github.com/SHI-Labs/NATTEN) is now [available as a pip package](https://www.shi-labs.com/natten/)!
+    * You can now install NATTEN with pre-compiled wheels, and start using it in seconds. 
+
 ### September 29, 2022
 * New preprint: [Dilated Neighborhood Attention Transformer](DiNAT.md).
-* [NA CUDA extension v0.13](NATTEN.md) released with dilation support!
-  * See [changelog](CHANGELOG.md).
+* [NA CUDA extension v0.13](https://github.com/SHI-Labs/NATTEN) released with dilation support!
 
 ### July 9, 2022
-* [NA CUDA extension v0.12](NATTEN.md) released.
+* [NA CUDA extension v0.12](https://github.com/SHI-Labs/NATTEN) released.
   * NA runs much more efficiently now, up to 40% faster and uses up to 25% less memory compared to Swin Transformer’s Shifted Window Self Attention.
   * Improved FP16 throughput.
   * Improved training speed and stability.
-  * See [changelog](CHANGELOG.md).
   
 ### May 12, 2022
-* [1-D Neighborhood Attention](NATTEN.md) support added!
+* [1-D Neighborhood Attention](https://github.com/SHI-Labs/NATTEN) support added!
 * Moved the kernel to `natten/` now, since there's a single version for all three tasks, and we're adding more features to the extension.
 
 ### April 30, 2022
-* [NA CUDA extension v0.11](NATTEN.md) released.
+* [NA CUDA extension v0.11](https://github.com/SHI-Labs/NATTEN) released.
   * It's faster in both training and inference, 
   * with a single version for all three tasks (no downstream-specific version)
-* [PyTorch implementation](NATTEN.md) released
+* [PyTorch implementation](https://github.com/SHI-Labs/NATTEN) released
   * Works both with and without CUDA, but not very efficient. Try to use the CUDA extension when possible.
-  * See [changelog](CHANGELOG.md).
-
-
-# Catalog
-- [x] Neighborhood Attention 1D (CUDA)
-- [x] Neighborhood Attention 2D (CUDA)
-- [ ] Neighborhood Attention 1D (PyTorch)
-- [x] Neighborhood Attention 2D (PyTorch)
-- [x] Dilation support
-- [ ] BFloat16 support (coming soon)
-- [ ] Zeros/Valid padding support (coming soon)
-- [ ] HuggingFace Demo
-
