--- classification/natten
@@ -1,1 +1,0 @@
-../natten/
