+++ classification/requirements.txt
@@ -2,7 +2,8 @@
 torch==1.11.0+cu113
 -f https://download.pytorch.org/whl/cu113/torch_stable.html
 torchvision==0.12.0+cu113
+-f https://www.shi-labs.com/natten/wheels/cu113/torch1.11/index.html
+natten
 git+https://github.com/rwightman/pytorch-image-models.git@9d6aad44f8fd32e89e5cca503efe3ada5071cc2a
-ninja==1.10.2.3
 fvcore==0.1.5.post20220305
 pyyaml==6.0
