+++ detection/requirements.txt
@@ -2,9 +2,10 @@
 torch==1.11.0+cu113
 -f https://download.pytorch.org/whl/cu113/torch_stable.html
 torchvision==0.12.0+cu113
+-f https://www.shi-labs.com/natten/wheels/cu113/torch1.11/index.html
+natten
 -f https://download.openmmlab.com/mmcv/dist/cu113/torch1.11.0/index.html
 mmcv-full==1.4.8
 mmdet==2.19.0
 git+https://github.com/rwightman/pytorch-image-models.git@9d6aad44f8fd32e89e5cca503efe3ada5071cc2a
-ninja==1.10.2.3
 fvcore==0.1.5.post20220305
