+++ segmentation/README.md
@@ -3,19 +3,22 @@
 ## Requirements
 Python 3.8 is strongly recommended.
 PyTorch version `1.11` is strongly recommended.
-Please read the instructions on how to set up and compile [NATTEN](../NATTEN.md) carefully before starting.
 For ease of use, you can just set up a new environment and run the following:
 ```shell
 pip3 install -r requirements.txt
 ```
-Similar to and because of [classification](../classification/README.md), semantic segmentation also depends on `timm`, `ninja`, 
+This will install the recommended torch and torchvision, our CUDA extension, and all other dependencies.
+
+Similar to and because of [classification](../classification/README.md), semantic segmentation also depends on `timm`, 
 and `fvcore`. Semantic Segmentation experiments were conducted with [mmsegmentation](https://github.com/open-mmlab/mmsegmentation/).
+Additionally, they depend on our extension, [NATTEN](https://github.com/SHI-Labs/NATTEN), which you can install 
+[by referring to our website](https://www.shi-labs.com/natten/). 
 The following are the recommended versions of these libraries and are strongly encouraged for reproducibility and speed:
 ```shell
 torch==1.11.0+cu113
 torchvision==0.12.0+cu113
+natten==0.14.1 # -f http://www.shi-labs.com/natten/wheels/cu113/torch1.11/index.html
 timm==0.5.0
-ninja==1.10.2.3
 fvcore==0.1.5.post20220305
 
 mmcv-full==1.4.8
